recipe_management_api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
drf-spectacular>=0.27
python-dotenv>=1.0
django-cors-headers>=4.3
